Don't duplicate 'DISTINCT' SQL keyword
`uniq`, when in the AR query context, adds the
`DISTINCT` SQL keyword to your queries. Explicitly
setting the `select` with a distinct and calling
`uniq` is redundant. Further, pluck is better when
only querying for a single column as it prevents
the full loading of AR model instances.

However, the `pluck` method was not available in
this context. I was getting this error in the Chunk
Helper specs:

    Failure/Error: GoogleSafeBrowsing::ChunkHelper.build_chunk_list.
    NoMethodError:
      undefined method `pluck' for #<Class:0x007fb178a50380>
    # ./lib/google_safe_browsing/chunk_helper.rb:22:in `block in build_chunk_list'
    # ./lib/google_safe_browsing/chunk_helper.rb:11:in `each'
    # ./lib/google_safe_browsing/chunk_helper.rb:11:in `build_chunk_list'
    # ./spec/chunk_helper_spec.rb:15:in `block (4 levels) in <top (required)>'